requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.9.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
